Exporting is still brittle for OTF traces and clustering.

+++ src/commrecord.h
@@ -9,7 +9,7 @@ class CommRecord
 public:
     CommRecord(unsigned int _s, unsigned long long int _st,
                unsigned int _r, unsigned long long int _rt,
-               unsigned int _size, unsigned int _tag,
+               unsigned long long _size, unsigned int _tag,
                unsigned int _group,
                unsigned long long int _request = 0);
 
@@ -18,7 +18,7 @@ public:
     unsigned int receiver;
     unsigned long long int recv_time;
 
-    unsigned int size;
+    unsigned long long size;
     unsigned int tag;
     unsigned int type;
     unsigned int group;
